Create README.md
Identifying a person or developing a person recognition system is one of the most challenging 
fields in computer vision. A key success factor in this field is the extraction of large amounts of data 
and classification of input images using deep learning algorithms. The application of deep learning 
algorithms to extract the most relevant features from an image and classify them into different 
categories produces outstanding results. By utilizing previously available data knowledge, CNN and 
Transfer learning algorithms assist in achieving the desired accuracy. As a result of transfer learning, 
previously acquired knowledge is utilized and training time is reduced. 
As deep learning and related technologies advance, the field of person recognition in computer 
vision is rapidly evolving. As a result of the use of CNNs, transfer learning, and other deep learning 
techniques, researchers and developers are making significant advances in the ability to accurately 
identify individuals in images, which have broad implications for security, surveillance, and 
personalized customer service.